fix: updated types of data, updated class with valid of object existence

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/UserController.java
@@ -29,7 +29,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable("id") Long userId) {
+    public User getUser(@PathVariable("id") int userId) {
         return userService.getUser(userId);
     }
 
@@ -40,25 +40,25 @@ public class UserController {
     }
 
     @PutMapping("/{id}/friends/{friendId}")
-    public Set<Long> addFriend(@PathVariable("id") Long userId,
-                               @PathVariable("friendId") Long friendId) {
+    public Set<Long> addFriend(@PathVariable("id") int userId,
+                               @PathVariable("friendId") int friendId) {
         return userService.addFriend(userId, friendId);
     }
 
     @DeleteMapping("/{id}/friends/{friendId}")
-    public Set<Long> deleteFriend(@PathVariable("id") Long userId,
-                                  @PathVariable("friendId") Long friendId) {
+    public Set<Long> deleteFriend(@PathVariable("id") int userId,
+                                  @PathVariable("friendId") int friendId) {
         return userService.deleteFriend(userId, friendId);
     }
 
     @GetMapping("/{id}/friends")
-    public List<User> getFriends(@PathVariable("id") Long userId) {
+    public List<User> getFriends(@PathVariable("id") int userId) {
         return userService.getFriends(userId);
     }
 
     @GetMapping("/{id}/friends/common/{otherId}")
-    public List<User> getCommonFriends(@PathVariable("id") Long userId,
-                                       @PathVariable("otherId") Long otherId) {
+    public List<User> getCommonFriends(@PathVariable("id") int userId,
+                                       @PathVariable("otherId") int otherId) {
         return userService.showCommonFriends(userId, otherId);
     }
 }
